Make ajax work properly on mobile
#276 On mobile, a modal popup is not used. However, many code references use the modal. This commit changes all jQuery selectors from using the modal as the parent, to using the form as the parent, which is present on both mobile and on desktop.

This also sets up the jQuery click handler on the “download” button to be the same on mobile or desktop. That fixes the problem where ajax calls were not being done on mobile when doing email verification, and instead users were being redirected to the ajax endpoint and seeing raw json.